Added process isolation to run computations safely in the background

+++ cdl/tests/multiprocess1_unit.py
@@ -26,37 +26,55 @@ import scipy.ndimage as spi
 import scipy.signal as sps
 from guiqwt.plot import ImageWindow
 from multiprocess import Pool
+from multiprocess.pool import AsyncResult
 from qtpy import QtWidgets as QW
 
 from cdl.env import execenv
 from cdl.tests.data import create_2d_random, create_test_image2
 from cdl.utils.qthelpers import create_progress_bar, qt_app_context
 
-SHOW = True  # Show test in GUI-based test launcher
+SHOW = False  # Do not test in GUI-based test launcher
+
+
+POOL: Pool = None
 
 
 class Worker:
     """Multiprocessing worker, to run long-running tasks in a separate process"""
 
     def __init__(self) -> None:
-        self.asyncresult = None
-        self.result = None
+        self.asyncresult: AsyncResult = None
+        self.result: Any = None
+
+    @staticmethod
+    def create_pool() -> None:
+        """Create multiprocessing pool"""
+        global POOL
+        # Create a pool with one process
+        POOL = Pool(processes=1)
+
+    @staticmethod
+    def terminate_pool() -> None:
+        """Terminate multiprocessing pool"""
+        global POOL
+        if POOL is not None:
+            POOL.terminate()
+            POOL.join()
+            POOL = None
 
     def run(self, func: Callable, args: tuple[Any]) -> None:
         """Run computation"""
         global POOL
-        assert self.asyncresult is None
+        assert POOL is not None
         self.asyncresult = POOL.apply_async(func, args)
 
     def terminate(self) -> None:
         """Terminate worker"""
-        global POOL
         # Terminate the process and stop the timer
-        POOL.terminate()
-        POOL.join()
+        self.terminate_pool()
         print("Computation cancelled!")
         # Recreate the pool for the next computation
-        POOL = Pool(processes=1)
+        self.create_pool()
 
     def is_computation_finished(self) -> bool:
         """Return True if computation is finished"""
@@ -71,7 +89,7 @@ class Worker:
 
 def test(iterations: int = 4) -> None:
     """Multiprocessing test"""
-    global POOL
+    Worker.create_pool()
     with qt_app_context(exec_loop=True):
         win = ImageWindow("Multiprocessing test", icon="datalab.svg", toolbar=True)
         win.resize(800, 600)
@@ -84,12 +102,15 @@ def test(iterations: int = 4) -> None:
                 progress.setValue(index)
                 progress.setLabelText(f"Computing {index}")
 
-                def func(data: np.ndarray, size: int) -> np.ndarray:
+                def func(data: np.ndarray, size: int, error: bool) -> np.ndarray:
+                    if error:
+                        raise ValueError("Test error")
                     return sps.medfilt(data, size) + create_2d_random(
                         data.shape[0], data.dtype
                     )
 
-                worker.run(func, (image.data, 3))
+                test_error = index == 2
+                worker.run(func, (image.data, 3, test_error))
                 while not worker.is_computation_finished():
                     QW.QApplication.processEvents()
                     time.sleep(0.1)
@@ -97,14 +118,15 @@ def test(iterations: int = 4) -> None:
                         worker.terminate()
                         break
                 if worker.is_computation_finished():
-                    image.data = worker.get_result()
+                    try:
+                        image.data = worker.get_result()
+                    except Exception as exc:
+                        print(f"Intercepted exception: {exc}")
                     win.get_plot().add_item(image.make_item())
                 else:
                     break
-        POOL.terminate()
-        POOL.join()
+        worker.terminate_pool()
 
 
 if __name__ == "__main__":
-    POOL = Pool(processes=1)
     test()
